requirements.txt
groq
python-fasthtml